requirements.txt
blinker==1.9.0
capstone==5.0.7
click==8.3.1
colorama==0.4.6
Flask==3.1.3
gunicorn==22.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
keystone-engine==0.9.2
MarkupSafe==3.0.3
Werkzeug==3.1.7
